refactor: merge OpenAIResponsesModel into OpenAIModel
Consolidates the OpenAI provider into a single class with `api: 'chat'`
(default) or `api: 'responses'`. Splits internals into per-API adapter
modules under src/models/openai/, mirroring the src/models/google/ layout.
Removes OpenAIResponsesModel as a separate export.
