+++ strands-ts/package.json
@@ -19,12 +19,8 @@
       "default": "./dist/src/models/anthropic.js"
     },
     "./models/openai": {
-      "types": "./dist/src/models/openai.d.ts",
-      "default": "./dist/src/models/openai.js"
-    },
-    "./models/openai-responses": {
-      "types": "./dist/src/models/openai-responses.d.ts",
-      "default": "./dist/src/models/openai-responses.js"
+      "types": "./dist/src/models/openai/index.d.ts",
+      "default": "./dist/src/models/openai/index.js"
     },
     "./models/bedrock": {
       "types": "./dist/src/models/bedrock.d.ts",
